Force Polymer 0.4.0
Now that 0.5.0 is out, bower was trying to pull in that
incompatible version because polymer internal deps have
circular dependencies with >=0.4.0, which bower interprets
as 0.5.0, and then pull in new versions, *with different
versioned circular dependencies*. Naturally causing conflicts.
Specifying resolution along not enough to solve it as bower
thought some of them were "unneccesary" and thus ignored.
So we had to fork the Polymer repos to fix the metadata, and
use those here and in all dependencies which use Polymer.
(the-graph,the-card,the-panel)