preview: Don't bundle a NoFlo browser build
Instead use https://noflo.org/noflo-browser/everything.html
as default URL for NoFlo browser projects.
This is just a build of NoFlo using grunt-noflo-browser which includes
a bunch of known stuff. Anyone can create similar builds, for instance
using https://github.com/noflo/noflo-browser-app as a template,
and point their IDE to it.
As long as the URL is served with AppCache this should also work offline,
after having fetched it the first time.